Completed move of some files
