+++ hedera-mirror-rest/server.js
@@ -35,7 +35,6 @@ const accounts = require('./accounts');
 const balances = require('./balances');
 const config = require('./config');
 const constants = require('./constants');
-const contracts = require('./contracts');
 const health = require('./health');
 const network = require('./network');
 const schedules = require('./schedules');
